requirements.txt
# Web Framework - FastAPI (keep for API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.31.0
psycopg2-binary>=2.9.9

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Validation and Serialization
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Security
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
alembic>=1.13.1

# Flask Web Application
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-login>=0.6.3
flask-cors>=4.0.0
werkzeug>=3.0.1
